Change LatLng properties from float to double

+++ src/Common/Common.Domain/Models/LatLng.cs
@@ -3,16 +3,16 @@
 namespace Common.Domain.Models;
 public sealed class LatLng : ValueObject
 {
-    public float Latitude { get; init; }
-    public float Longitude { get; init; }
+    public double Latitude { get; init; }
+    public double Longitude { get; init; }
 
-    private LatLng(float latitude, float longitude)
+    private LatLng(double latitude, double longitude)
     {
         Latitude = latitude;
         Longitude = longitude;
     }
 
-    public static LatLng Create(float latitude, float longitude) => new(latitude, longitude);
+    public static LatLng Create(double latitude, double longitude) => new(latitude, longitude);
 
     protected override IEnumerable<object?> GetEqualityComponents()
     {
